Allow rules with no "alwaysUpdated" artifacts.
It is currently not allowed to have a rule whose artifacts all have the
"alwaysUpdated" property set to false, but there does not seem to be a
reason for that. Allowing this actually enables us to have rules that
are run unconditionally, which does not seem to be possible at the
moment (without forcing a "physical" timestamp check).

Task-number: QBS-262
Change-Id: I0ad57d0df2af4605e0c713925ab43a428af7ffa0
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>