Try harder not to store inconsistent build graphs.
In particular, try to prevent that nodes of a product that is no longer
part of the project still appear as children of nodes in other products.
While it is not entirely certain that this solves the linked bug (as I
don't know how to trigger it), it is now hopefully no longer possible
that such an inconsistent state is actually written to disk (that is, if
the problem still exists, our sanity checks should now catch it when
storing the build graph, not when loading).

Task-number: QBS-682
Change-Id: Id6d55246ac3de014e6b917abbd6aed1505769968
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>