Check internalVersion instead of product.version for creating symlinks.
This allows users to remove the version number from a shared library
filename by setting internalVersion to undefined. Without this change an
exception is thrown if shouldCreateSymlinks is not also explicitly set
to false when this is done.

Later, internalVersion should probably be renamed to something more
sensible ("fileNameVersion", "versionSuffix", etc.) and made public and
documented.

Change-Id: I4a3dcf6d871440250887bc3ec8abd377542b903b
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>