API: Make sure newly added files are sorted.
A popular convention is to keep the entries of a "files" list sorted. If
we do not respect this, the "add files" feature becomes much less
useful, as users will have to edit the file afterwards.
So make sure we do not break such an ordering, if it exists (but do not
change the order of existing files, as we cannot know if the user wants
that).
No control over the way how to sort is currently given to the user.

Change-Id: Iec6f0e8a6df85399c906f7f1090ddb76b5f3046e
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>