API: Do not export disabled sub-projects.
As opposed to disabled products or groups, we know next to nothing
about these projects, possibly not even their name (in the case of
a non-existing file conditionally referenced by a SubProject item).
Let's not make them available via the API then, as no useful
operations on them are possible.

Change-Id: Iad6055b65f2a7329739b1e4c88f9626a26354df5
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>