fix evaluation scope of module properties
Consider module A that depends on module B.
Module properties of B were evaluated in the wrong scope.
The comment in evaluatorscriptclass.cpp states correctly
"Own properties of module instances must not have the instance
itself in the scope.", but the corresponding check was slightly
wrong. We must check whether the item that the property belongs to
is a module instance or not.

Task-number: QBS-515
Change-Id: Ic96b83fbca8b9ea783165a15b7c8398640a4dd84
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>