introduce cpp.runtimeLibrary
This property determines the type of the used runtime library.
Accepted values are \c{"static"} and \c{"dynamic"}.
If this property is set to \c{undefined}, then the default runtime
library of the toolchain is used.
For MSVC the default value is "dynamic".
At the moment this property is only functional for MSVC.
Support for other toolchains may be added later.

Change-Id: I6795e89280d057cdb31967ef5c114ef60b3a98be
Task-number: QBS-570
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>