Introduce convenience syntax for dependency on all product instances.
In certain contexts (e.g. Android packages), one typically wants to
depend on all products of a certain name, regardless of the profile for
which it was built. This is now possible by specifying an empty
"profiles" list in the Depends item (which does not otherwise make
sense). Without this feature, users would have to duplicate the list
that they already provided in the product, which is obviously not
desirable.

Change-Id: If3460bc45355689917ab5ecac587ae6083931dd3
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@theqtcompany.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>