support installation of directory trees
Introduce the string property qbs.installSourceBase that specifies the
base directory of the local file that is going to be installed.
  The default is the directory path of the current file to be installed,
relative to the product's source directory. This ensures that the current
behavior is not changed.

In this example

Group {
    files: ["data/**/*.png"]
    qbs.install: true
    qbs.installDir: "content"
    qbs.installSourceBase: "data"
}

the directory structure of the PNG files below "data" is preserved and
is installed in a "content" subdirectory.

Change-Id: Ic0c7166b07da77a8ae7eea98aa345079f74e17e9
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>