fix look-up of scalar properties
The property map of a product used to contain all modules of the
dependency tree as direct children. The same was true for the
product's Item. This setup was done in
ModuleLoader::createAdditionalModuleInstancesInProduct.
The whole approach led to problems when retrieving module values
from modules deeper in the dependency hierarchy. Basic idea of this
fix is to make ProjectResolver::resolveModule recursively walk the
module dependency tree instead of trying to gather information on the
root level.

Change-Id: Icf1325272f90d7924437f41ca6f403d7264ce84c
Task-number: QBS-726
Task-number: QBS-706
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>