fix scalar property lookup some more
Let there be a product P that depends on module A. Module A depends on
module B.A sets B.foo to 1. Default value of B.foo is 0.  The value of
product.moduleProperty("B", "foo") is 1.

Now add the direct dependency B to P. P does not overwrite the default
value of B.foo. The overwritten value of B.foo from A takes
precedence.

This semantics change is needed to make sure that adding/removing
modules to a product does not change property values in surprising
ways.

Task-number: QBS-736
Change-Id: I73152a24649bc670a9075e82998ef2b0d8709a84
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>