Don't evaluate properties that will never be seen.
During project resolving, we can make use of our knowledge
of how the PropertyFinder works, skipping evaluation for
"shadowed" properties.
In the long term, we might want to get rid of the nested
modules map altogether and replace it with a "flat" map
(with special support for mergeable list values).

Task-number: QBS-746
Change-Id: I75f62239daadce1f8826444a2e3b3a51d581df98
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>