Don't recursively copy directory artifacts.
If the input artifact is a directory, the concatentation of the
installDir and the input artifact's last path component is created as a
directory path, rather than that directory being recursively copied into
the installDir. This is necessary because recursively copying directory
artifacts is problematic and hard to properly track dependencies for,
since artifacts may be installed in an undefined order.

Change-Id: I2e408302f0015695885049b1d9db5d3bde152948
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>