Improve change tracking.
A long time ago, we introduced two categories of changed products in the
change tracker: Products with "simple changes" such as a new file tag or
a different module property value, and those with "severe changes" that were
very likely to result in a need to re-build artifacts, such as a change
in a prepare script. The latter's build data was simply thrown away and
re-generated, whereas for the former we carefully applied the necessary
changes to the build data "manually" with the idea to minimize the amount
of rebuilds.
However, some time later we introduced the concept of "rescue data",
which eliminated unnecessary rebuilds for the "severe changes" case as
well. In fact, it turns out that this approach is actually much more
robust and reliable in addition to having the simpler implementation. So
we now get rid of the hand-crafted special handling for "simple
changes", as it is just inferior.

Task-number: QBS-724
Task-number: QBS-791
Change-Id: Ia06ca4ce5697a13cdd48d002acdab1cc2904a044
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>