fuzzy tester: Cache information about build results.
We remember the information whether a clean build succeeded. This
enables us to skip potentially expensive verification builds as well as
finding a usable start commit.
Note: This can lead to false positives or negatives if qbs or the build
environment have changed in a way that leads to a formerly succeeding
build failing or vice versa; however, this should be rare and easily
spottable.

Change-Id: I690f7a01fc1174fa50214775b0faa3b550168963
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>