GCC module: Add option for fine-tuning relinking behavior.
Before this patch, we considered a dynamic library as sufficiently
changed to require relinking if the number of public symbols had changed.
However, this can lead to behavior like the following:
    - Open the Qt Creator project and add a call to qDebug() somewhere
      in the Core plugin.
    - Rebuild and watch in amazement as all other plugins get relinked.
The reason for that is that there was no qDebug() call in this library,
so adding one introduces a new undefined public symbol into the binary.
This patch allows it to ignore such changes. The property introduced for
this purpose is a string as opposed to a boolean, because it could
conceivably be extended later.
Pushed to 1.4 because the high annoyance factor makes the current behavior
bug-like.

Task-number: QBS-101
Change-Id: I671716c9ddb56ce895eaad25ef243339bef092ee
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@theqtcompany.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>