Clean up handling of {Project,Product}.qbsSearchPaths.
The code was redundant and generally much more complicated than it
needed to be.
Note: This change introduces a (backward-compatible) behavior change, as
search paths set in products and child projects no longer overwrite the
respective value from the parent project. The reason for the original
decision was probably to allow exotic use cases like finding a module of
the same name at a different location, but this is still possible, as
the paths deeper in the inheritance tree take precedence.

Change-Id: If2e9b55c2ab6e1b22e57b1b4943ab6b52540a744
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>