fix property lookup
Merge the modules of the same name in the module hierarchy of a
product into one. This way we get correct dependency resolution
between properties.
Remove the awful hacks that we have introduced to solve QBS-736 and
QBS-706.
Instead of a hierarchy of QVariantMaps representing all module
values we store a flat map that contains the final (potentially
merged) module values. This leads to a significant speed up in
project loading and rule execution.

Task-number: QBS-736
Change-Id: I4dc6ce9566118a9f2bdc4536b1449164899db0e8
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>