Check for the item type when trying to load a module.
Situation without this patch: If a product is located in a directory
that is in the module search path, we will happily load the product file
as a module, and dependencies on this product will not have the desired
effect.
Situation with this patch: When trying to load what could be a module,
we check the item type and reject the file as a module if the type is
not "Module". As a result, the file will be loaded as a product, as
expected.

Task-number: QBS-806
Change-Id: I9510cf61b8075fc74b49ea334dcba94bb63b626a
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>