Fix several issues with path handling in the TypeScript module.
- QBS-412 comment is removed and replaced with the appropriate solution.
- The intermediate directory is no longer a concatenation of absolute
  paths, which was broken on Windows due to the colon character.
- Properly return zero artifacts from a product which has no TypeScript
  source files (avoids a "null pointer" dereference).
- Document the convolution around what's now referred to as the
  "compiledIntermediateDir", which is basically a workaround for QBS-5.
- Try to preserve directory structure of input TypeScript/js files.
- Adapt autotests.

There are still some issues, namely that the TypeScript compiler
builds external modules from source if it finds them in the path,
making it non-trivial to preempt the list of output files. The best
solution might be to patch the TypeScript compiler itself (upstream, of
course) to print a listing of output files that will be generated for a
particular invocation, in combination with --noEmit (dry run).

Change-Id: I6be47c0eed72a6196880264d0eb18bad9eea1ea9
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>