Initialize qbs.commonRunEnvironment from the current environment.
We used to start with an empty map, which leads to modules'
setupRunEnvironment scripts overwriting the environment instead of
enriching it. For instance, when an application built with qbs was
run from Qt Creator, it did not get the environment specified by
the user, but only what qbs' scripts set up.

Change-Id: I13bcba8f051d773cc135dc7bb91440a88b2c18c4
Reviewed-by: Jake Petroules <jake.petroules@petroules.com>