Solve the Java dependency tracking issue once and for all.
This approach relies on the newly introduced ability of Modules to
contain Groups, which may add additional source files to a Product.

In the case of the Java module, the module includes a group which adds
some sources to build a helper tool on-demand for each product that
builds Java source code. This deviates from the previous approach of
supplying a single jar for the entire qbs distribution and is thus
somewhat less efficient, but is not expected to be a critical issue in
the short term.

To start, it solves two critical issues: the build-time dependency on
javac, which complicates the qbs build process unnecessarily, and the
bootstrap class path warning due to targeting Java 1.6 with newer JDKs.

Change-Id: I533214a5d37fd69a4d8bfcf0db36bc7dae821d5f
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>