Make qbs.architecture an optional property.
The value "undefined" now indicates an architecture-independent target
like the CLR or JVM.

The cpp module's copy of the architecture property is not required
unconditionally either, because there exist implementations that compile
C/C++ to architecture-independent targets (Emscripten). However it
remains required for GCC and MSVC toolchains where a specific
architecture is always targeted.

Change-Id: I0810e71bd561d3b9bc1f83d70329ee78d129e9db
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>