Make it possible to overwrite properties conditionally.
If a product or higher-level module wants to set the property of a
(different) module only if some condition is fulfilled and leave it
unchanged otherwise, it would currently need to duplicate the value
from the module prototype. Obviously, this is not feasible for anything
but the most trivial definitions, and even then it easily introduces
maintenance problems.

Change-Id: If33c93a25dc84603fe95704e8815aa6e3aa1e0e4
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>