Introduce BlackboxOutputArtifactTracker.
Since many tools unfortunately don't offer dry-run and/or output
listing options, this provides a reusable unit used to compile
to a temporary directory and list the resulting file contents.

This will be used for Interface Builder tools (asset catalogs,
XIBs/NIBs, storyboards) as well as qdoc (which can be fixed
upstream but still requires this for older versions) and probably
Java.

Change-Id: I1d40ab068ac396c87cf825afc0d5eedf4dbf90da
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>