Fix performance regression when importing JS files.
A side effect of the bugfix commit 33c9a2721b was that two imports of
the same file were no longer considered the same if the import location
differed. This results in a considerable slowdown due to more cache
misses, so restore the old behvior in this regard.
Also, JSImport's qHash() and operator==() are now consistent.

Change-Id: I23acd665a6329ec1ab132cb34657f0b05357093f
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>