Add an item flag designating the item as a "module prefix".
We use this ad-hoc term to describe items that represent
the non-last parts of multi-segment module property bindings.
For instance: In the binding "Qt.core.version", "Qt" is
a module prefix, whereas "core" is a module instance.
This enables us to get rid of the ugly data structure that
kept information about which item values belonged to which
parent.

Change-Id: I4100f9623775afb94a28c414f03623f212f1ed50
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>