Add an option to require transformers to generate all declared outputs.
This ensures that a successful build cannot complete without all output
artifacts of all rules and transformers in the project existing (unless
they are marked alwaysUpdated: false).

This is undesirable to enable by default due to the potential overhead,
but can be a useful debugging tool and so is provided as an option.

Change-Id: I46ee49e28cc06d5aff9a3cfd520f929fef4de3f8
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>