Fix false positives in property declaration check.
If we attach values to properties of a conditionally pulled-in module
inside a Properties block, that binding gets inserted into the
surrounding module. Without this patch, the property declaration check
will then emit a diagnostic for that property if the condition of the
respective Depends item is false. We fix this by skipping the check for
bindings that were not originally present in the surrounding item.
Note: This means we will not catch some bindings anymore that are
actually wrong. It could probably be done, but would involve an
inordinate amount of book-keeping, plus additional evaluations of
Properties conditions.

Task-number: QBS-854
Change-Id: I75c2af00fdfb97c9db2de2241f701dd213a401da
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>