Fix cleaning of rescuable artifacts.
The following sequence was problematic:
    1) qbs build someModuleProperty:true
    2) qbs resolve someModuleProperty:false
    3) qbs clean
Step 2) causes the artifact objects to get removed from the build graph
and their relevant information stored in the list of rescuable artifact
data, so that on the next build the files will either get re-created as
artifacts or removed from the disk, depending on whether the changed
property causes their transformer command to change.
The problem was that the ArtifactCleaner only looked at the build graph
and was not aware of rescuable artifact data. With this patch, we go
through the list of rescuable artifacts and remove the entries from that
list as well as from the disk.

Change-Id: I05348fa1e92268c58f75074ea99371e39cc662f4
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>