Use destinationDirectoryForResource for all ib module output paths.
This fixes an issue where asset catalogs compiled their resources into
the wrong directory due to the ib module reading properties of the
bundle module from the dependency in its own scope, rather than the
product's scope.

This could be triggered by, for example, setting the consoleApplication
property in a product to true (which causes bundle.isBundle to be false)
while also explicitly setting bundle.isBundle to true.

Change-Id: Ie014d42d4d8bef5e0b253c0f5a79036b4791ed55
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>