ProjectResolver: Remove check for product vs module name conflicts.
It offers no value, as the condition of a product having the same name
as one of its dependencies is already caught by the check for cyclic
dependencies. Note that the "module" does not need to be an actual
module, but could also be a product, in which case the error message
probably was more confusing than helpful anyway.

Change-Id: I62bd3980ae377dfb5acb8f34039cb47add7a1c31
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>