Change tracking: Be more sensitive to environment changes.
Contrary to what the old code assumed, it is not enough to compare the
values of environment variables that were used while resolving the
project: A command run by a transformer might be sensitive to
environment changes as well (e.g. gcc to CPLUS_INCLUDE_PATH).
Note: This does not mean we will rebuild everything when anything in the
environment changes; it means that *if* a file is rebuilt, its command
will see the new environment instead of the old one.

Change-Id: I7916b8c40f1440564034ec3317c549ebd4e988a0
Reviewed-by: Jake Petroules <jake.petroules@theqtcompany.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>