Qt Core module: Make the lrelease rule optionally a multiplex rule.
lrelease takes a number of ts files and produces a qm output, so it is
technically a multiplex rule. However, normally users have one ts file
per language and want a 1:1 mapping. So we need to make this
configurable as not to force people to create one product per ts file.

Change-Id: I1b14eca6b52fd18a792e1ba26e3fe55db371ae3b
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>