Disallow modules where one's name equals the other's first component.
That's just asking for trouble. For instance, in the existing
TestLanguage::modules() autotest, debugging revealed that the expression
"dummy.deep.moat.depth" on the rhs of a property in the product was
leading to an item value property "deep" being created in the module
"dummy", which is highly unexpected and could lead to serious problems
in a more complex project.
While it would perhaps be possible to support this kind of stuff
properly, the whole idea seems dubious to begin with and I do not think
it is worth the non-negligible effort that would have to be spent there.

Change-Id: Ie938c99d24990f23e8dd73f70026f94f3bc5e4a0
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>