ModuleLoader: Proper evaluation of product conditions.
The following cases did not work correctly due to ordering problems:
    - Product conditions depending on the presence of another product
      whose condition in turn depends on the presence of a third
      product.
    - Product conditions depending on the results of Probes.
We solve this by splitting up product item handling even more. The
second stage now handles dependencies. Afterwards, we run a third stage
that executes probes and evaluates the product condition. For the third
stage, the order of products matters in the sense that all products that
are dependencies are handled before the ones that depend on them.

Change-Id: I1d436fb811796d4375294aa92f1cd088dbbe30de
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>