Fix unwanted "concurrency" in Executor.
Script engine evaluation is interrupted once a second to check for
cancel requests. At this time, jobs can finish, leading to the
onJobFinished() slot being called while we are in the middle of
executeRuleNode(). Possible effects:
    - The job that just finished was the only currently running one,
      and its transformer outputs have no parents. Then the Executor
      will think it's done, and higher-level code will free the
      evaluation context (and with it the engine). When script execution
      continues after the current round of events processing, a null
      pointer or freed memory will be accessed.
    - The job's transformer outputs do have parent artifacts. In this
      case, we are suddenly and most unexpectedly in executeRuleNode()
      twice, with the potential for all kinds of stunning behavior.
Fix the problem by checking in the onJobFinished() slot whether the
rules evaluation context is currently active and delaying slot execution
if it is.

Task-number: QBS-932
Change-Id: I03ca4f4ad844357e49d5acaddff066395a7f95cf
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>