Fix change tracking for prepare scripts using JS imports.
Comparing the old and new source code of a rule is not enough if part of
the code is factored out into a JavaScript helper file and the code
there has changed. So we now take the filestamp of imports into account
when deciding whether a prepare script will have to be re-run.
Note that this will lead to some false positives, as we do not easily
get at the information which import is actually used in which prepare
script.

Change-Id: Iab5be90fccddff339e9d4fe36aff6a5c208ad9c2
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>