Simplify JS import code
Remove the 'scope' parameter from ScriptEngine::importProgram.
It was either a) empty or b) the same as the target value.
In case a) it wasn't used.
In case b) it was needed to have the scope of the RulesEvaluationContext
(e.g. the JS constructors Command and JavaScriptCommand) available in
rules.

Interestingly, the scope of RulesEvaluationContext is also pushed to the
ScriptEngine's current context. But QScriptContext::pushContext in
importFile creates a new context with a fresh scope chain.
Restore the scope chain of the former current context after pushContext.

Consequently, magic variables like "path" and "filePath" are not available
within imported JS files anymore. Those properties were reflecting the
importing module anyway, which is wrong within the imported JS file.

Change-Id: I5899d4bb931717a462e8d02ca1d0e56e39431e62
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>