Fix handling of flat dSYM output and greatly expand test coverage.
If flat dSYM output was requested, the generated dwarf files would be
placed next to the associated executable or bundle wrapper with a .dSYM
extension appended. Now, the dwarf files will be placed next to the
associated executable (which itself might be within a bundle) with a
.dwarf extension.

This follows the default behavior of dsymutil when no output file is
explicitly specified.

Change-Id: Iba71d4fe43b7415d33a244728ed8bce515270583
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>