Do extensive validation on the value of qbs.toolchain.
qbs now enforces the order of the items in the toolchain list, ensures
that there are no duplicates, and ensures that all implied values are
present in the list (for example, clang implies llvm implies gcc).

Users can now rely on appropriate values always being present (like
llvm, or gcc), and on the order of the toolchain list.

Change-Id: I9abcee9265a8ac9539639128bb4b624353c155ec
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>