Automatically determine the compiler's architecture.
Also, do extensive validation on the result by ensuring that the arch
specified by the user is what which the compiler will actually generate
when invoked, by constructing an appropriate argument list and doing a
dump of the compiler's macros.

This ensures a user cannot specify a bogus architecture which is not
actually the one the compiler will generate. Furthermore, compiler
probes will no longer generate the wrong value for qbs.architecture
for ARM-family architectures, since the triple returned by
$(CC) -dumpmachine does not include the subarch which needs to be passed
to the compiler.

Users will also be prevented from subverting proper operation by
attempting to pass raw ABI-related flags in the flags properties.

Change-Id: I2aed8e0afefdf7e5a4112183f2f9ea96b1e22b51
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>