Revisit importing JS source code
Import code without access to the activation object.
Other script engines do not allow easy access to the activation object.
The new import code evaluates the code to be imported in an anonymous
function that returns an object containing all variables and functions
that are declared on the top level.

E.g. importing the following snippet
    var greeting = 'hey!';
    function upperCaseGreeting() { return "HEY!"; }
is evaluated as
    (function () {
        var greeting = 'hey!';
        function upperCaseGreeting() { return "HEY!"; }
        return {
            greeting : greeting,
            upperCaseGreeting : upperCaseGreeting
        }
    })()

This has the following advantages:
  - Imported JS files do not see variables/functions anymore that
    were defined in already imported JS files.
  - No specialities of QScriptEngine are used.
    Porting to another JS engine becomes easier.
  - The clearing of the import cache before evaluation of every rule
    can be removed which improves the performance of rule execution
    by 28 % (according to qbs_benchmarker).

Change-Id: I2f635c3375a137a96c27abf89d04a27fdf2c3f23
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>