Fix edge case in change tracking.
If we do some change that causes a transformer to be re-created, the
information about e.g. project properties accessed in commands got lost
and subsequent changes to these properties would then not get detected
anymore. This means we have to store that type of information in our
rescuable artifact data too.

Change-Id: I8d23aa4471019670c060d8a51fee5dc835ed07c9
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>