Respect the conditions of Export items.
This did not happen before, because module conditions are only evaluated
when loading the module file, of which there is none in the case of
Export items.
Note: The implementation assumes that the condition of an Export item
has the semantics "use or ignore this Export item". The other possible
meaning would be to use the condition as the condition of the resulting
module instance, but that immediately leads to weirdness: What does it
mean for the dependency to a product if said product is enabled, but its
Export item is not? What about conflicting conditions in the Export
items of an inheritance chain -- do they get logically "and"ed? And so
on.

Change-Id: I08d7d93d30b8a53b42a317cb3391924e4666d169
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>