qbs build: Move build config properties into a dedicated module.
Qt Creator will do the same in order to be able to build plugins that
are not part of the main source tree. This means the qbs build and
install settings can no longer be controlled by properties in a higher-
level project file, but need to come from a module.

Change-Id: I20c0a4538395c8ee838b33f35be84ee59f601f90
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>