Fix null pointer accesses.
We used to call certain member functions on an object that could be
null. While that was never allowed, it did not cause problems so far
because these functions checked their "this" pointer and returned
immediately if it was null. With GCC 6, such constructs no longer work.
Measures taken:
    - Fix the null pointer accesses by allocating the respective
      objects earlier (in case they must not be null) or checking for a
      null value at the calling site (in case they can legitimately be
      null).
    - Rewrite the aforementioned member functions so that they will
      no longer silently tolerate undefined behavior with other
      compilers than GCC 6.

Task-number: QBS-974
Change-Id: If07ddc5c0e536d36c19311357f871554aa692ad6
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Ilya Lyubimov <villytiger@gmail.com>