Allow Java dependency tracking to work without actually compiling.
This uses the Java compiler APIs to traverse the AST of the Java code
to find type elements and get their corresponding binary names (i.e.
base names of the generated class files). Because this method does not
require the source files to be compiled, a complete* list of output
artifacts can be retrieved even if some dependencies are missing, as
long as the source is well-formed.

*No attempt is made to handle annotation processors which generate
additional source files. Qbs has not previously supported this anyways.
It may do so in the future.

Change-Id: I9e4231071618c020f1087c455d436cdb14c1f8e9
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>