Fix edge case in change tracking.
If a product has two or more rules with the same set of input and output
tags (this is possible in the case of no input tags), then it could
happen that one of them was compared to itself instead of to the other
one. This is the reason for the "propertyChanges" autotest being flaky
lately.

Change-Id: Iabd3f87086b50be7694562b0fde574901c31cf42
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>