Executor: Do not assert on condition that can actually happen.
We asserted on the condition that a node's product must be in the list
of products to build, meaning that it needs to belong to the set of
user-selected products or one of their dependencies. However, there are
at least two ways in which one can trigger this condition to be false:
1) The user provides a faulty project in which a "Depends" item is
missing, but the respective artifact from the other product is still
found (e.g. a generated header file via a project-global include
path).
2) The project is actually okay, but our C++ scanner erroneously adds
dependencies from a different product, e.g. because it does not know
about #ifdefs.

Instead of the assertion, we now simply mark the respective node as
built and continue. Ideally, we'd log a warning, but we cannot do that
because of point 2) above.

Change-Id: I3549d732dea5cde84d1019132580a8e051c9db11
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>