Do the OS X to macOS rename.
[ChangeLog] The "osx" value in qbs.hostOS and qbs.targetOS has been
replaced with "macos". For backwards compatibility, qbs.hostOS will
still contain "osx" in addition to "macos", and an error will be printed
if qbs.targetOS contains "osx" but not "macos". Specifying "macos"
without "osx" is allowed. These checks are subject to be removed in a
future version of qbs, so update your hostOS and targetOS checks
accordingly.

[ChangeLog] cpp.minimumOsxVersion has been deprecated and replaced with
cpp.minimumMacosVersion. cpp.minimumOsxVersion is subject to be removed
in a future version of qbs, so update your projects accordingly.

Change-Id: I479891829dff6eb6750cb2a04e1395f085896f63
Reviewed-by: Jake Petroules <jake.petroules@qt.io>