ModuleLoader: Add dependency cycle check
This is necessary, because otherwise we might silently skip some
products, which can lead to unexpected results.
There is no overhead in the normal code path, because the fact that we
have a cycle can be detected for free by the product sorter.
The check in the ProjectResolver is still needed, because wildcard
dependencies are only resolved later.
Added autotests to cover more cases.

Change-Id: Ie3bf1db5bbb3eeb95d40f5139d294aac6ad840c1
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>